+++ src/algorithms/operators/mod.rs
@@ -7,6 +7,7 @@ pub(crate) mod complementer;
 pub(crate) mod difference;
 pub(crate) mod disjoint_union;
 pub(crate) mod intersection;
+pub(crate) mod is_same_graph;
 pub(crate) mod simplify;
 pub(crate) mod union;
 
@@ -14,5 +15,6 @@ pub use complementer::complementer;
 pub use difference::difference;
 pub use disjoint_union::{disjoint_union, disjoint_union_many};
 pub use intersection::intersection;
+pub use is_same_graph::is_same_graph;
 pub use simplify::simplify;
 pub use union::union;